application/requirements.txt
Django==3.0
channels==2.4.0
django-bootstrap4==2.2.0
psycopg2-binary==2.8.6
djangorestframework==3.11.1
djoser==2.0.5
